refactor: Split NIP-20 implementation code into separate plugins.
