+++ src/nips/nip-0020-command-results/plugins/index.test.ts
@@ -1,0 +1,15 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview Tests for commandResults().
+ */
+
+import { setupTestHubAndClient } from '../../../test/setup-test-hub-and-client';
+import { commandResults } from '.';
+
+describe('commandResults()', () => {
+  it('should be a plugin', () => {
+    setupTestHubAndClient(commandResults);
+  });
+});
